requirements.txt
aliyun-python-sdk-core-v3==2.13.11
amqp==2.5.2
billiard==3.6.2.0
celery==4.4.0
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==2.1.7
django-ckeditor==5.9.0
django-filter==2.2.0
django-js-asset==1.2.2
django-oauth-toolkit==1.2.0
djangorestframework==3.11.0
drf-yasg==1.17.0
entrypoints==0.3
flake8==3.7.9
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jmespath==0.9.4
kombu==4.6.7
MarkupSafe==1.1.1
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.1
Pillow==7.0.0
pycodestyle==2.5.0
pycryptodome==3.9.7
pyflakes==2.1.1
PyMySQL==0.9.3
pyparsing==2.4.6
pytz==2019.3
redis==3.4.1
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
uritemplate==3.0.1
urllib3==1.25.8
vine==1.3.0
zipp==1.2.0
